Updated kalloc.c

+++ mytest.c
@@ -34,7 +34,7 @@ int main() {
     //char* temp4 = (char*)mmap(12288,4096,PROT_READ|PROT_WRITE,MAP_POPULATE|MAP_ANONYMOUS,-1,0); //this will occur pagefault.
     //if(temp4 == 0) exit();
     //printf(1,"Free Page : %d!\n",freemem());
-    munmap(temp),munmap(temp2);
+    munmap((uint)temp),munmap((uint)temp2);
     printf(1,"Free Page : %d!\n",freemem());
     exit();
 }
